Add WIP Traits, Generics, etc.
This isn't quite what I was going for;
I wanted to make this more generic so
as not to rely on u32 type but to be able
to swap out different types for HP.

It looks like Trait Aliases is
something that will get me closer to that
goal.

I think a lesson I learned here is that if
you set up some generic bounds to a certain
Trait(s), you will only have access to those
trait fns or types and not the properites of the
struct that is implementing those traits.

In other words, bound an impl to either a
concrete struct or to a trait, expecting
both in your implementation doesn't make sense.